[refactor] 디렉토리 변경 및 loggin level 변경
